requirements.txt
aiohttp==3.6.2
discord.py==1.3.3
aiosqlite==0.13.0
discord==1.0.1
docker_py==1.10.6
python-dotenv==0.13.0
